backend/requirements.txt
django
django-cors-headers
djangorestframework
django-channels
channels-redis
deepface
djangorestframework-simplejwt
daphne
opencv-python
requests